Captions/requirements.txt
flask==2.3.3
flask-cors==4.0.0
openai-whisper==20231117
yt-dlp==2023.12.30
datasets==2.15.0
nltk==3.8.1
torch==2.1.0
torchaudio==2.1.0
numpy==1.24.3